requirements.txt
fastapi
uvicorn
python-multipart
google-genai
google-cloud-storage
pydantic
qrcode
pillow
